[gstreamer] Acquire STREAM_LOCK on the sink pad before flushing libav decoder. Fixes JB#9140
FLUSH_START is not serialized with the buffers. It gets sent to us while we are in the middle
of pushing a frame.

We simply flush libav bufers which leads to freeing the data while we are still copying the frame.
This leads to a crash due to invalid memory being accessed,

We need to serialize the event by acquiring the stream lock in order to prevent the race condition
from happening. A better option would be to set a flag but it will be hard to synchronize
(Taking easy path for now).