[gstreamer] use gst_av_vdec::mutex to protect frame also in convert_frame(). Fixes JB#9140
We cannot use the STREAM_LOCK because if we are in PAUSED state and we try to acquire
that lock then we will deadlock because the pad is already blocked due to the sink
prerolling.

It makes more sense to use gst_av_vdec::mutex because convert_frame() manipulates
the frame data by copying so it's considered part of decoding.
This will also prevent freeing the buffer data when we flush libav.

The other alternative would be decoupling the source and sink pads completely by
using a GstTask which really complicates the element for nothing.