[gstreamer] delay setting srcpad caps until we have decoded the first few samples. Fixes JB#9156
Some audio streams are weird. Both aacparse and qtdemux detect the sample rate to be 22050Hz
yet libav gives us back 44,1KHz.

When we calculate the buffer duration we use the sample rate
obtained from libav (44.1K in this case) but we set the buffer
caps based on the input caps (sample rate = 22050 Hz). This simply
causes a gap in the stream and the audio to skip.

We simply delay setting the caps until we have decoded the first few
samples and are ready to push the buffer downstream. This allows us to
use the info from libav which gets filled only after the decoding.